# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ResearchRequest {
  /** The energy research query */
  query: string;
  /** Maximum number of sources to search (default 5) */
  maxSources?: number;
}

export type ResearchJobResponseStatus =
  (typeof ResearchJobResponseStatus)[keyof typeof ResearchJobResponseStatus];

export const ResearchJobResponseStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface ResearchJobResponse {
  jobId: string;
  status: ResearchJobResponseStatus;
  message: string;
}

export type ResearchJobStatusStatus =
  (typeof ResearchJobStatusStatus)[keyof typeof ResearchJobStatusStatus];

export const ResearchJobStatusStatus = {
  pending: "pending",
  running: "running",
  completed: "completed",
  failed: "failed",
} as const;

export interface Reference {
  url: string;
  title: string;
  snippet?: string;
}

export interface ResearchReport {
  id: string;
  query: string;
  title: string;
  introduction: string;
  keyInsights: string[];
  conclusion: string;
  references: Reference[];
  createdAt: string;
  sources?: string[];
}

export interface ResearchJobStatus {
  jobId: string;
  status: ResearchJobStatusStatus;
  query: string;
  /** Current progress message */
  progress?: string;
  report?: ResearchReport;
  error?: string;
  startedAt?: string;
  completedAt?: string;
}

export interface ResearchReportSummary {
  id: string;
  query: string;
  title: string;
  createdAt: string;
  sourceCount: number;
}

export interface ResearchHistoryResponse {
  reports: ResearchReportSummary[];
  total: number;
}

export interface SearchRequest {
  query: string;
  topK?: number;
}

export interface SearchResult {
  reportId: string;
  query: string;
  title: string;
  score: number;
  snippet: string;
  createdAt: string;
}

export interface SearchResponse {
  results: SearchResult[];
  total: number;
}

export interface ErrorResponse {
  error: string;
  detail?: string;
}