# lib/api-zod/src/generated/types/reference.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Reference {
  url: string;
  title: string;
  snippet?: string;
}