# lib/api-zod/src/generated/types/researchHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchReportSummary } from "./researchReportSummary";

export interface ResearchHistoryResponse {
  reports: ResearchReportSummary[];
  total: number;
}