# lib/api-zod/src/generated/types/researchJobResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchJobResponseStatus } from "./researchJobResponseStatus";

export interface ResearchJobResponse {
  jobId: string;
  status: ResearchJobResponseStatus;
  message: string;
}