# lib/api-zod/src/generated/types/researchJobStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ResearchJobStatusStatus } from "./researchJobStatusStatus";
import type { ResearchReport } from "./researchReport";

export interface ResearchJobStatus {
  jobId: string;
  status: ResearchJobStatusStatus;
  query: string;
  /** Current progress message */
  progress?: string;
  report?: ResearchReport;
  error?: string;
  startedAt?: string;
  completedAt?: string;
}