# lib/api-zod/src/generated/types/researchReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { Reference } from "./reference";

export interface ResearchReport {
  id: string;
  query: string;
  title: string;
  introduction: string;
  keyInsights: string[];
  conclusion: string;
  references: Reference[];
  createdAt: string;
  sources?: string[];
}