# lib/api-zod/src/generated/types/researchReportSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchReportSummary {
  id: string;
  query: string;
  title: string;
  createdAt: string;
  sourceCount: number;
}