# lib/api-zod/src/generated/types/researchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchRequest {
  /** The energy research query */
  query: string;
  /** Maximum number of sources to search (default 5) */
  maxSources?: number;
}