# lib/api-zod/src/generated/types/searchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchRequest {
  query: string;
  topK?: number;
}