# lib/api-zod/src/generated/types/searchResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { SearchResult } from "./searchResult";

export interface SearchResponse {
  results: SearchResult[];
  total: number;
}