# lib/api-zod/src/generated/types/searchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Autonomous Energy Researcher Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface SearchResult {
  reportId: string;
  query: string;
  title: string;
  score: number;
  snippet: string;
  createdAt: string;
}